- continue implementation of bet, customer and market microservices - started implementation of dao and service layer for bet service - will implement kafka handlers for events and commands when everything is ready

+++ common-domain/src/main/java/net/skycomposer/betting/common/domain/dto/customer/WalletData.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class WalletData {
     private String walletId;
-    private Integer balance;
+    private BigDecimal balance;
 }
